pyproject.toml
[project]
name = "sqlpilot"
version = "0.1.0"
description = "A LLM Agent based SQL optimization and verification platform."
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pymysql>=1.1.0",
    "aiomysql>=0.2.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.2",
    "openai>=1.0.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
